rearrage my files to separate concerns
