+++ src/Components/posts/postsJSFiles/Post.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { withRouter } from 'react-router-dom'
-import {checkToken} from "../checkToken"
+import {checkToken} from "../../checkToken"
 
 const api = 'https://teamy-api.herokuapp.com/api/v1/posts';
 // eslint-disable-next-line react-hooks/rules-of-hooks
@@ -37,10 +37,14 @@ class Post extends Component {
     render() {
         const individualPost = this.state.post;
         return (
-            <div className="postis-container">
+            <div className="postid-container">
                 <h1>{`${individualPost.firstname}'s post`}</h1>
-                <div>
+                <div className="individual-post-dp">
                     <img src={`https://res.cloudinary.com/amaka01/image/upload/v1609578087/cg3mtemxniugidu73ewn.jpg`} alt="dp"/>
+                </div>
+                <div className="individual-post-div">
+                    <p>{individualPost.title}</p>
+                    <p>{individualPost.article}</p>
                 </div>
             </div>
         )
